refactor(core): move `injectArgs` into `util.ts`
This commit moves the `injectArgs` function that is used in multiple
places into `util.ts` file. It should help avoid circular dependency
issues when the function is updated or used in other places.

PR Closes #36479
